百分比位置修改

+++ app/src/main/java/com/growatt/chargingpile/util/MyUtil.java
@@ -622,7 +622,7 @@ public class MyUtil {
     }
 
 
-    public static double divide(float v1, float v2, int scale) {
+    public static double divide(double v1, double v2, int scale) {
         BigDecimal b1 = new BigDecimal(Double.toString(v1));
         BigDecimal b2 = new BigDecimal(v2 + "");
         return b1.divide(b2, scale, BigDecimal.ROUND_HALF_UP).doubleValue();
